+++ examples/plugin_package/src/pyingestkit_demo_jobs/__init__.py
@@ -1,0 +1,3 @@
+from .local_file import DemoLocalFileJob, job
+
+__all__ = ["DemoLocalFileJob", "job"]